1.0

+++ ShopApi/Repositories/UserRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Models;
+using Models.Interfaces;
 using Repositories.Interfaces;
 
 namespace Repositories
@@ -13,6 +14,22 @@ namespace Repositories
             this.db = db;
         }
 
+        public User? Retrieve(string email)
+        {
+            IQueryable<User> users = db.Users.Where(u=>
+                u.Email == email
+            );
+
+            User? user = null;
+            foreach (User u in users)
+            {
+                user = u;
+            }
+
+            return user is null ? null : user;
+
+        }
+
         public async Task<User?> CreateAsync(User data)
         {
             User user = (await db.Users.AddAsync(data)).Entity;
@@ -22,6 +39,11 @@ namespace Repositories
         public async Task<bool> DeleteAsync(int id)
         {
             User? user = db.Users.Find(id);
+
+            if(user is null)
+            {
+                return false;
+            }
 
             db.Users.Remove(user);
             return await db.SaveChangesAsync() == 1 ? true : false;
